requirements.txt
absl-py==1.4.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
ConfigArgParse==1.7
dotted==0.1.8
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
idna==3.4
importlib-metadata==6.8.0
joblib==1.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
numpy==1.24.4
oauthlib==3.2.2
Pillow==10.0.0
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
timeout-decorator==0.5.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.1
trimesh==3.23.3
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
zipp==3.16.2
